refactor(core): break grants import cycles via grants_resolve leaf
Extract feat_catalog and grants_resolve so feats/proficiencies no longer
lazy-import character_build; expose parse_character_class and load helpers.
