+++ core/feats.py
@@ -1,21 +1,32 @@
-"""Черты персонажа: загрузка, применение, требования, видимость, описания."""
+"""Черты персонажа: применение, требования, видимость, описания.
+
+Каталог YAML и чистое чтение grants — ``core.feat_catalog``.
+"""
 
 import re
 from dataclasses import dataclass, field
-from pathlib import Path
 from typing import Any
 
-from core.catalog_loader import load_catalog
 from core.classes import character_has_spellcasting
-from core.grants import (
-    normalize_armor_token,
-    proficiency_tokens_and_skills_from_grant,
+from core.feat_catalog import (
+    FeatGrant,
+    get_feat_expertise_ids,
+    get_feat_language_ids,
+    get_feat_proficiency_grants,
+    get_feat_save_proficiencies,
+    get_feat_skill_ids,
+    get_race_feat_grants,
+    iter_feat_grants,
+    load_feat,
+    load_feats,
+    race_feat_step_required,
+    resolve_feat_grants,
 )
+from core.grants import normalize_armor_token
 from core.grants_context import CreationContext
+from core.grants_resolve import resolve_grants_for_context
 from core.hp_bonus import HpBonusSource, hit_point_bonus_amount
 from core.types import StatMap
-
-FEATS_FILE = Path("database/progression/feats.yaml")
 
 _BENEFITS_MARKER = re.compile(
     r"(?:,\s*)?(?:и\s+)?(?:вы\s+)?(?:дающие\s+)?"
@@ -40,17 +51,8 @@ _PROFICIENCY_GRANT_TYPES = frozenset(
 
 
 # ============================================================================
-# Dataclasses и загрузка
+# Dataclasses
 # ============================================================================
-
-
-@dataclass(frozen=True)
-class FeatGrant:
-    """Выборная черта от расы/подрасы."""
-
-    count: int
-    from_list: str
-    source: str
 
 
 @dataclass(frozen=True)
@@ -69,32 +71,6 @@ class FeatRequirementContext:
     level: int = 1
     has_spellcasting: bool = False
     skills: list[str] = field(default_factory=list)
-
-
-def _load_feats_yaml() -> dict[str, Any]:
-    """Загрузить feats из YAML."""
-    return load_catalog(FEATS_FILE, "feats")
-
-
-def load_feats() -> list[dict[str, Any]]:
-    """Список всех черт."""
-    result: list[dict[str, Any]] = []
-    for feat_id, info in _load_feats_yaml().items():
-        if isinstance(info, dict):
-            entry = dict(info)
-            entry["id"] = feat_id
-            result.append(entry)
-    return result
-
-
-def load_feat(feat_id: str) -> dict[str, Any]:
-    """Данные одной черты."""
-    info = _load_feats_yaml().get(feat_id, {})
-    if isinstance(info, dict):
-        entry = dict(info)
-        entry["id"] = feat_id
-        return entry
-    return {"id": feat_id}
 
 
 # ============================================================================
@@ -232,107 +208,6 @@ def resolve_feat_ability_bonuses(
     return bonuses
 
 
-def resolve_feat_grants(
-    feat_id: str, choices: dict[str, Any] | None = None
-) -> tuple[list[str], list[str], list[str], list[str]]:
-    """Владения из черты с учётом подвыборов."""
-    feat = load_feat(feat_id)
-    choices = choices or {}
-    weapons: list[str] = []
-    armors: list[str] = []
-    tools: list[str] = []
-    skills: list[str] = []
-    raw_grants = feat.get("grants", [])
-    if not isinstance(raw_grants, list):
-        return weapons, armors, tools, skills
-    for grant in raw_grants:
-        if not isinstance(grant, dict):
-            continue
-        w, a, t, s = proficiency_tokens_and_skills_from_grant(grant, choices)
-        weapons.extend(w)
-        armors.extend(a)
-        tools.extend(t)
-        skills.extend(s)
-    return weapons, armors, tools, skills
-
-
-def get_feat_skill_ids(
-    feat_ids: list[str],
-    feat_choices: dict[str, dict[str, Any]] | None = None,
-) -> list[str]:
-    """Навыки из выбранных черт."""
-    feat_choices = feat_choices or {}
-    skills: list[str] = []
-    for feat_id in feat_ids:
-        choices = feat_choices.get(feat_id, {})
-        _, _, _, s = resolve_feat_grants(feat_id, choices)
-        skills.extend(s)
-    return skills
-
-
-def get_feat_language_ids(
-    feat_ids: list[str],
-    feat_choices: dict[str, dict[str, Any]] | None = None,
-) -> list[str]:
-    """Языки из черт (linguist)."""
-    feat_choices = feat_choices or {}
-    langs: list[str] = []
-    for feat_id in feat_ids:
-        choices = feat_choices.get(feat_id, {})
-        raw = choices.get("languages", [])
-        if isinstance(raw, list):
-            langs.extend(str(lang) for lang in raw)
-    return langs
-
-
-def get_feat_expertise_ids(
-    feat_ids: list[str],
-    feat_choices: dict[str, dict[str, Any]] | None = None,
-) -> list[str]:
-    """Навыки с экспертным владением из черт (skill_expert)."""
-    from core.skills import PHB_SKILL_IDS
-
-    feat_choices = feat_choices or {}
-    skills: list[str] = []
-    for feat_id in feat_ids:
-        raw = feat_choices.get(feat_id, {}).get("expertise", [])
-        if isinstance(raw, list):
-            for item_id in raw:
-                sid = str(item_id)
-                if sid in PHB_SKILL_IDS and sid not in skills:
-                    skills.append(sid)
-    return skills
-
-
-def get_feat_save_proficiencies(
-    feat_ids: list[str],
-    feat_choices: dict[str, dict[str, Any]] | None = None,
-) -> list[str]:
-    """Владение спасбросками из черт (Resilient)."""
-    feat_choices = feat_choices or {}
-    saves: list[str] = []
-    for feat_id in feat_ids:
-        feat = load_feat(feat_id)
-        choices = feat_choices.get(feat_id, {})
-        raw_grants = feat.get("grants", [])
-        if not isinstance(raw_grants, list):
-            continue
-        for grant in raw_grants:
-            if not isinstance(grant, dict):
-                continue
-            if grant.get("type") != "save_proficiency":
-                continue
-            if grant.get("choice"):
-                picked = choices.get("ability")
-                if isinstance(picked, str) and picked not in saves:
-                    saves.append(picked)
-            else:
-                target = grant.get("ability")
-                if isinstance(target, str) and target not in saves:
-                    saves.append(target)
-    return saves
-
-
 def apply_feats_to_stats(
     stats: StatMap,
     feat_ids: list[str],
@@ -400,41 +275,19 @@ def apply_feat_grants_to_character(
     )
 
 
-def get_feat_proficiency_grants(
-    feat_id: str,
-    choices: dict[str, Any] | None = None,
-) -> tuple[list[str], list[str], list[str], list[str]]:
-    """Владения из черты: (weapons, armors, tools, skills)."""
-    return resolve_feat_grants(feat_id, choices)
-
-
-def _iter_feat_grants(feat_ids: list[str]) -> list[dict[str, Any]]:
-    """Grants из выбранных черт."""
-    grants: list[dict[str, Any]] = []
-    for feat_id in feat_ids:
-        feat = load_feat(feat_id)
-        raw_grants = feat.get("grants", [])
-        if not isinstance(raw_grants, list):
-            continue
-        for grant in raw_grants:
-            if isinstance(grant, dict):
-                grants.append(grant)
-    return grants
-
-
 def has_non_light_dual_wield(feat_ids: list[str]) -> bool:
     """Черта «Использование двух оружий» — сражение без свойства «лёгкое»."""
     return any(
         grant.get("type") == "dual_wielder"
         and grant.get("non_light_dual_wield")
-        for grant in _iter_feat_grants(feat_ids)
+        for grant in iter_feat_grants(feat_ids)
     )
 
 
 def dual_wielder_ac_bonus_from_feats(feat_ids: list[str]) -> int:
     """Бонус КД из grant dual_wielder (обычно +1)."""
     bonus = 0
-    for grant in _iter_feat_grants(feat_ids):
+    for grant in iter_feat_grants(feat_ids):
         if grant.get("type") != "dual_wielder":
             continue
         amount = grant.get("ac_bonus", 0)
@@ -482,8 +335,6 @@ def creation_known_for_feat_picks(
     level: int,
 ) -> tuple[list[str], list[str], list[str]]:
     """Навыки, инструменты и токены оружия до подвыборов внутри черты."""
-    from core.character_build import resolve_grants_for_context
-
     ctx = _creation_context(
         race_id, subrace_id, background_id, class_id, subclass_id, level
     )
@@ -513,8 +364,6 @@ def build_feat_selection_context(
     Опциональные ``skills`` / ``weapon_tokens`` / ``tool_tokens`` дополняют
     владения расы, класса и предыстории (например, от уже выбранных черт).
     """
-    from core.character_build import resolve_grants_for_context
-
     ctx = _creation_context(
         race_id,
         subrace_id,
@@ -692,33 +541,6 @@ def feat_visible_for_selection(
 # ============================================================================
 # Требования черт
 # ============================================================================
-
-
-def get_race_feat_grants(
-    race_id: str, subrace_id: str | None = None
-) -> list[FeatGrant]:
-    """Слоты выбора черты из grants расы/подрасы."""
-    from core.grants import grants_of_type
-    from core.races import collect_race_grants
-
-    result: list[FeatGrant] = []
-    for grant in grants_of_type(
-        collect_race_grants(race_id, subrace_id), "feat"
-    ):
-        count = int(grant.get("count", 1))
-        from_list = str(grant.get("from", "all"))
-        source = "subrace" if subrace_id else "race"
-        result.append(
-            FeatGrant(count=count, from_list=from_list, source=source)
-        )
-    return result
-
-
-def race_feat_step_required(
-    race_id: str, subrace_id: str | None = None
-) -> bool:
-    """Нужен ли шаг выбора черты при создании."""
-    return bool(get_race_feat_grants(race_id, subrace_id))
 
 
 def _armor_requirement_met(
